Moved run() out of __init__ and into autonicer
